feat: Some fixes in generator
Attempt to convert policies of more services uncovered certain issues in
the generator.  Unfortunately Neutron is again doing what it is famous for
- crazy stuff. It has custome policy checks: OwnerCheck (that dynamically
fetches attributes of the parent/owner resources) and FieldCheck which is
extracting certain field or also fetches network resource. It is not really
possible to generate proper policy for that until a way of acessing such
additional information is being found. There are few ways but the most
promising seems to be implementing a tiny http proxy app that exposes only
required attributes of required resources by looking them in the database
directly. This data is very dynamic and cannot be simply loaded into the
OPA as data.