fix: Rename member_id to member for glance (#31)
Glance is using key name transformation when accessing the ImageTarget.
Bypassing oslo.policy causes that target dictionary contains "member"
attribute, while the policy is willing to access "member_id". Since
there is no reasonable easy way of dealing with that on the oslo.policy
plugin side rather rename the attribute in the policy.
While similar renaming is done for "owner" and "id" those does not seem
to be necessary (id is not used anywhere in the policy and owner is
being renamed to the project_id which is already existing in the target
dict).