feat: Fix tests generation with custom policy
When a custom policy file is passed it may contain additional rules.
Till now we were relying on the strict order of rules to appear so that
when a rule relies on other rule it was already previously converted.
This change splits policy generation and test generation into 2 separate
runs (the first one for converting rules and the second one generating
tests for the converted rules). This change once again faces
maintainability questions with all methods in the single module,
therefore split it into smaller pieces.