feat: Unwrap `"foo":%(a.b)s` similarly to oslo_policy
oslo_policy first tries to access property directly and when it is
missing goes into the lookup. Barbican explicitly flattens the dict
resulting in structure like `"input": {"target":
{"target.container.acl": ...}}`. Therefore during OPA policy generation
we need to explicitly allow that type of access as well.

This change results in a certain rework of test data handling in AND and
OR checks (funny, that it actually helped to simplify the code). Add
unittests for AND and OR checks test data generation.

NOTE: There is still tricky case remaining where oslo_policy would also
do special handling when path element is a list. This can not be really
addressed easily.