feat: Drop internal attributes from Neutron in opa_filter (#50)
When opa_filter check is invoked fields that are not intended to be
returned to the user must be also removed from the output. Neutron is
applying some a bit weird logic that we should repeat - get attributes
of the record and ensure that the attribute contains is_visible = True
on the controller level.