feat: Use urllib in opa check directly
In production we see sporadic timeouts. Analysis shows that on the
system level OPA responses are received well before the timeout (10ms vs
2sec timeout) but some services (namely Nova and Cinder) still fall with
the timeout. We suspect that there might be something in the eventlet
area or the old version of requests lib that we got stick to. Try to
replace requests usage with the urllib and see whether it helps.